requirements.txt
flask
flask_login
flask_sqlalchemy
python-dotenv
Flask-JWT-Extended
requests
pytest
pytest-flask
